Fix double addition bug

+++ zapit-backend/src/main/java/com/shaunwah/zapitbackend/model/LocationData.java
@@ -5,18 +5,16 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class LocationData {
     @JsonIgnore
     private Long id;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
 
-    public LocationData(BigDecimal latitude, BigDecimal longitude) {
+    public LocationData(Double latitude, Double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
